Address comments

+++ internal/api/snapshot_backup_schedule/snapshot_backup_schedule.go
@@ -1,8 +1,8 @@
 package snapshot_backup_schedule
 
 type SnapshotBackupSchedule struct {
-	Interval      int      `json:"interval"`
-	Retention     int      `json:"retention"`
+	Interval      int64    `json:"interval"`
+	Retention     int64    `json:"retention"`
 	StartTime     string   `json:"startTime"`
 	CopyToRegions []string `json:"copyToRegions"`
 }
